fix: add text only chapter support for local files

+++ Shared/Sources/Local/LocalFileManager.swift
@@ -64,38 +64,44 @@ extension LocalFileManager {
         }
 
         // find image entries (pages)
-        let imageEntries = archive
+        let pageEntries = archive
             .filter { entry in
                 let lastPathComponent = entry.path.lastPathComponent()
                 guard !lastPathComponent.hasPrefix(".") else {
                     return false
                 }
-                let ext = entry.path.lowercased().pathExtension()
+                let ext = entry.path.pathExtension().lowercased()
+                if ext == "txt" {
+                    return !entry.path.hasSuffix("desc.txt")
+                }
                 return Self.allowedImageExtensions.contains(ext)
             }
             .sorted {
                 $0.path.localizedStandardCompare($1.path) == .orderedAscending
             }
 
-        guard !imageEntries.isEmpty else {
+        guard !pageEntries.isEmpty else {
             return nil
         }
 
         // extract the first three images for preview
-        let previewImages = imageEntries.prefix(3).compactMap { entry -> PlatformImage? in
-            var imageData = Data()
-            do {
-                _ = try archive.extract(
-                    entry,
-                    consumer: { data in
-                        imageData.append(data)
-                    }
-                )
-                return PlatformImage(data: imageData)
-            } catch {
-                return nil
+        let previewImages = pageEntries
+            .filter { !$0.path.lowercased().hasSuffix("txt") }
+            .prefix(3)
+            .compactMap { entry -> PlatformImage? in
+                var imageData = Data()
+                do {
+                    _ = try archive.extract(
+                        entry,
+                        consumer: { data in
+                            imageData.append(data)
+                        }
+                    )
+                    return PlatformImage(data: imageData)
+                } catch {
+                    return nil
+                }
             }
-        }
 
         let fileType = switch pathExtension {
             case "cbz": LocalFileType.cbz
@@ -107,7 +113,7 @@ extension LocalFileManager {
             url: url,
             previewImages: previewImages,
             name: url.lastPathComponent,
-            pageCount: imageEntries.count,
+            pageCount: pageEntries.count,
             fileType: fileType
         )
     }
@@ -119,38 +125,83 @@ extension LocalFileManager {
         guard let cbzPath = await LocalFileDataManager.shared.fetchChapterArchivePath(mangaId: mangaId, chapterId: chapterId)
         else { return [] }
 
-        // read zip file
         let documentsDir = FileManager.default.documentDirectory
-        let archiveUrl = documentsDir.appendingPathComponent(cbzPath)
+        let archiveURL = documentsDir.appendingPathComponent(cbzPath)
+        return readPages(from: archiveURL)
+    }
+
+    // read pages from an archive file
+    nonisolated func readPages(from archiveURL: URL) -> [AidokuRunner.Page] {
         let archive: Archive
         do {
-            archive = try Archive(url: archiveUrl, accessMode: .read)
+            archive = try Archive(url: archiveURL, accessMode: .read)
         } catch {
+            LogManager.logger.error("Failed to read archive: \(error)")
             return []
         }
 
-        // find image entries (pages)
-        return archive
+        var descriptionFiles: [Entry] = []
+
+        var pages = archive
             .filter { entry in
+                // ignore hidden files
                 let lastPathComponent = entry.path.lastPathComponent()
                 guard !lastPathComponent.hasPrefix(".") else {
                     return false
                 }
-                let ext = entry.path.lowercased().pathExtension()
-                return Self.allowedImageExtensions.contains(ext) || ext == "txt"
+                // ensure extension is allowed
+                let ext = entry.path.pathExtension().lowercased()
+                if ext == "txt" {
+                    if entry.path.hasSuffix("desc.txt") {
+                        descriptionFiles.append(entry)
+                        return false
+                    }
+                    return true
+                }
+                return Self.allowedImageExtensions.contains(ext)
             }
             // sort by file name
             .sorted {
                 $0.path.localizedStandardCompare($1.path) == .orderedAscending
             }
             .map { entry in
-                AidokuRunner.Page(content: .zipFile(url: archiveUrl, filePath: entry.path))
+                AidokuRunner.Page(content: .zipFile(url: archiveURL, filePath: entry.path))
             }
+
+        for entry in descriptionFiles {
+            guard
+                let index = entry.path
+                    .lastPathComponent()
+                    .split(separator: ".", maxSplits: 1)
+                    .first
+                    .flatMap({ Int($0) }),
+                index > 0,
+                index <= pages.count
+            else { break }
+
+            do {
+                var descriptionData = Data()
+                _ = try archive.extract(
+                    entry,
+                    consumer: { data in
+                        descriptionData.append(data)
+                    }
+                )
+                pages[index - 1].hasDescription = true
+                pages[index - 1].description = String(data: descriptionData, encoding: .utf8)
+            } catch {
+                LogManager.logger.error("Failed to extract page description text from archive: \(error)")
+                continue
+            }
+        }
+
+        return pages
     }
 }
 
 extension LocalFileManager {
     // add a new file to the local files source
+    // swiftlint:disable:next cyclomatic_complexity
     func uploadFile(
         from url: URL,
         // if the file already exists, skip uploading it to avoid duplicates
@@ -216,20 +267,23 @@ extension LocalFileManager {
         }
 
         // find image entries (pages)
-        let imageEntries = archive
+        let pageEntries = archive
             .filter { entry in
                 let lastPathComponent = entry.path.lastPathComponent()
                 guard !lastPathComponent.hasPrefix(".") else {
                     return false
                 }
-                let ext = entry.path.lowercased().pathExtension()
+                let ext = entry.path.pathExtension().lowercased()
+                if ext == "txt" {
+                    return !entry.path.hasSuffix("desc.txt")
+                }
                 return Self.allowedImageExtensions.contains(ext)
             }
             .sorted {
                 $0.path.localizedStandardCompare($1.path) == .orderedAscending
             }
 
-        guard !imageEntries.isEmpty else {
+        guard !pageEntries.isEmpty else {
             throw LocalFileManagerError.noImagesFound
         }
 
@@ -311,18 +365,22 @@ extension LocalFileManager {
             }
         } else if mangaId == nil {
             // copy first page image to use as cover image
-            let firstImageEntry = imageEntries.first!
-            let coverExt = (firstImageEntry.path as NSString).pathExtension
-            let coverFileName = "cover.\(coverExt)"
-            let newCoverURL = mangaFolder.appendingPathComponent(coverFileName)
-            do {
-                if newCoverURL.exists {
-                    try? fileManager.removeItem(at: newCoverURL)
+            let firstImageEntry = pageEntries.first(where: { !$0.path.lowercased().hasSuffix("txt") })
+            if let firstImageEntry {
+                let coverExt = (firstImageEntry.path as NSString).pathExtension
+                let coverFileName = "cover.\(coverExt)"
+                let newCoverURL = mangaFolder.appendingPathComponent(coverFileName)
+                do {
+                    if newCoverURL.exists {
+                        try? fileManager.removeItem(at: newCoverURL)
+                    }
+                    _ = try archive.extract(firstImageEntry, to: newCoverURL)
+                    coverURL = newCoverURL
+                } catch {
+                    throw LocalFileManagerError.fileCopyFailed
                 }
-                _ = try archive.extract(firstImageEntry, to: newCoverURL)
-                coverURL = newCoverURL
-            } catch {
-                throw LocalFileManagerError.fileCopyFailed
+            } else {
+                coverURL = nil
             }
         } else {
             coverURL = nil
